+++ cotests/__init__.py
@@ -1,4 +1,4 @@
 from .batch import bench_batch, test_batch
 from .case.case import CoTestCase
-from .group import CoTestGroup, test_groups
+from .group.group import CoTestGroup, test_groups
 from .module import test_module
