--- cotests/group/__init__.py
@@ -1,1 +1,0 @@
-from .group import CoTestGroup, test_groups
